healthcare/requirements.txt
DBUtils==2.0.1
Flask==2.0.1
urllib3==1.26.4
psycopg2-binary==2.8.6
pyyaml
gevent==21.1.2
flask_cors==3.0.10
pydantic==1.8.1
flask-restful==0.3.9
gunicorn==20.1.0
Flask-Script==1.0.0
flasgger==0.9.5

